# src/docs/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: My API
  description: API documentation for my app
  version: 1.0.0
paths:
...
  /api/v1/transactions/update/{id}:
    post:
      summary: Update a transaction by ID
      operationId: updateTransactionById
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        200:
          description: Transaction updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Transaction updated successfully
  /api/v1/transactions/delete/{id}:
    post:
      summary: Delete a transaction by ID
      operationId: deleteTransactionById
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        200:
          description: Transaction deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Transaction deleted successfully
...
  /api/v1/categories/delete/{id}:
    post:
      summary: Delete a category by ID
      operationId: deleteCategoryById
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        200:
          description: Category deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Category deleted successfully
...